refactor(arduino): extract _sendAckAndFlush helper to reduce duplication
- Add _sendAckAndFlush() to encapsulate ACK + flush sequence
- Replace 6 duplicate ACK+flush blocks in dispatch() with helper calls
- Reduces code duplication and centralizes ACK logic
- Fix test_coverage_gaps.cpp to use int16_t for PID handler

All host tests pass.

+++ openwrt-library-arduino/tests/test_coverage_gaps.cpp
@@ -503,9 +503,9 @@ static void test_process_poll_response_short_payload() {
 }
 
 static bool g_process_async_called = false;
-static int g_process_async_pid = -1;
+static int16_t g_process_async_pid = -1;
 
-static void process_async_handler(int pid) {
+static void process_async_handler(int16_t pid) {
   g_process_async_called = true;
   g_process_async_pid = pid;
 }
